Add --log-level and --verbose-level command line options

+++ src/log/Logger.h
@@ -38,7 +38,7 @@ namespace logger {
 
         static void setLogLevel(int level);
 
-        static void setVerboseLevel(unsigned short level);
+        static void setVerboseLevel(int level);
 
         static void logToFile(const std::string& logFile);
